requirements.txt
# Stock Prediction Application Requirements
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Stock data fetching
yfinance>=0.2.18

# Machine learning
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Web application
streamlit>=1.28.0

# Database connectivity
psycopg2-binary>=2.9.0

# Additional utilities
python-dotenv>=1.0.0
plotly>=5.15.0
